a storage table dbcontext sample

+++ EFCore.Azure.StorageTable.Sample/Program.cs
@@ -1,29 +1,38 @@
 ﻿using Microsoft.EntityFrameworkCore;
 
-using var db = new BloggingContext();
+if (false)
+{
+    using var db = new BloggingContext();
 
-// Note: This sample requires the database to be created before running.
-Console.WriteLine($"Database path: {db.DbPath}.");
+    // Note: This sample requires the database to be created before running.
+    Console.WriteLine($"Database path: {db.DbPath}.");
 
+    // Create
+    Console.WriteLine("Inserting a new blog");
+    db.Add(new Blog { Url = "http://blogs.msdn.com/adonet" });
+    await db.SaveChangesAsync();
+
+    // Read
+    Console.WriteLine("Querying for a blog");
+    var blog = await db.Blogs
+        .OrderBy(b => b.BlogId)
+        .FirstAsync();
+
+    // Update
+    Console.WriteLine("Updating the blog and adding a post");
+    blog.Url = "https://devblogs.microsoft.com/dotnet";
+    blog.Posts.Add(
+        new Post { Title = "Hello World", Content = "I wrote an app using EF Core!" });
+    await db.SaveChangesAsync();
+
+    // Delete
+    Console.WriteLine("Delete the blog");
+    db.Remove(blog);
+    await db.SaveChangesAsync();
+}
+
+using var table = new BookContext();
 // Create
-Console.WriteLine("Inserting a new blog");
-db.Add(new Blog { Url = "http://blogs.msdn.com/adonet" });
-await db.SaveChangesAsync();
-
-// Read
-Console.WriteLine("Querying for a blog");
-var blog = await db.Blogs
-    .OrderBy(b => b.BlogId)
-    .FirstAsync();
-
-// Update
-Console.WriteLine("Updating the blog and adding a post");
-blog.Url = "https://devblogs.microsoft.com/dotnet";
-blog.Posts.Add(
-    new Post { Title = "Hello World", Content = "I wrote an app using EF Core!" });
-await db.SaveChangesAsync();
-
-// Delete
-Console.WriteLine("Delete the blog");
-db.Remove(blog);
-await db.SaveChangesAsync();
+Console.WriteLine("Inserting a new book");
+table.Add(new Book { Id = 1, Title = "Book 1" });
+await table.SaveChangesAsync();
